+ enable project nullable types check small refactoring
