+++ src/Dex.Pagination/Data/PageFilter.cs
@@ -1,4 +1,4 @@
-﻿namespace Dex.Pagination
+﻿namespace Dex.Pagination.Data
 {
     public class PageFilter : IPageFilter
     {
