+++ src/Dex.Pagination/Data/SortParam.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Dex.Pagination.Data
+{
+    public class SortParam : ISortParam
+    {
+        public string FieldName { get; set; } = String.Empty;
+        public bool IsDesc { get; set; }
+    }
+}
