--- src/Dex.Pagination/SortParam.cs
@@ -1,8 +1,0 @@
-﻿namespace Dex.Pagination
-{
-    public class SortParam : ISortParam
-    {
-        public string FieldName { get; set; }
-        public bool IsDesc { get; set; }
-    }
-}
